add fix/use client on root layout
