+++ app/layout.tsx
@@ -1,9 +1,9 @@
-"use client";
 import Navbar from "@/components/Navbar";
 import "./globals.css";
 import { Roboto } from "next/font/google";
 import Sidebar from "@/components/Sidebar";
 import SidebarProvider from "@/context/SidebarProvider";
+import Theme from "@/context/Theme";
 
 const roboto = Roboto({
   weight: ["100", "300", "400", "500", "700", "900"],
@@ -15,30 +15,6 @@ export const metadata = {
   description: "Test",
 };
 
-import { createTheme, ThemeProvider } from "@mui/material/styles";
-
-const theme = createTheme({
-  components: {
-    MuiAppBar: {
-      styleOverrides: {
-        root: {
-          background: "#fff", // Set the background color to transparent
-          boxShadow: "none", // Remove the box shadow
-        },
-      },
-    },
-    MuiButton: {
-      styleOverrides: {
-        root: {
-          "&:hover": {
-            background: "none", // Set the hover background color to transparent
-          },
-        },
-      },
-    },
-  },
-});
-
 export default function RootLayout({
   children,
 }: {
@@ -47,7 +23,7 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={roboto.className}>
-        <ThemeProvider theme={theme}>
+        <Theme>
           <SidebarProvider>
             <header className="border-b fixed z-40 top-0 left-0 w-full">
               <Navbar />
@@ -56,7 +32,7 @@ export default function RootLayout({
               <Sidebar>{children}</Sidebar>
             </main>
           </SidebarProvider>
-        </ThemeProvider>
+        </Theme>
       </body>
     </html>
   );
